requirements.txt
alembic==1.8.0
bcrypt==3.2.2
blinker==1.4
cffi==1.15.0
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.2
jedi==0.17.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
mypy==0.961
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysql-connector-python==8.0.29
parso==0.7.1
pdfkit==1.0.0
pluggy==1.0.0
protobuf==4.21.2
pycparser==2.21
pyls-mypy @ git+https://github.com/tomv564/pyls-mypy.git@3b105b491a42eec304f19e6ed6da1053d12958c6
PyMySQL==1.0.2
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
SQLAlchemy==1.4.37
tomli==2.0.1
typing_extensions==4.2.0
ujson==5.3.0
Werkzeug==2.1.2
WTForms==3.0.1
